requirements.txt
yaaredis==3.0.0
celery==5.3.6
kombu==5.3.4
redis==3.5.3
flower==2.0.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
# psycopg2==2.9.9
celery-redbeat==2.1.1